Revert "Ajout des salles dans Engagement + modification des nouvelles + changement de photo"
