+++ content/pages/a-propos/nos-collections/nos-collections-speciales.mdx
@@ -22,8 +22,7 @@ import SciencesHumaines from '../../images/sciences-humaines-Perrault-Pouchkine.
 import Theologie from '../../images/theologie-hore-virginis-marie.webp'
 import SansImage from '../../images/collection-sans-image.webp'
 import SansImage2 from '../../images/collection-sans-image2.webp'
-import Droit from '../../images/collection-speciale-droit.jpg'
-import LSH from '../../images/collection-speciale-lettres-sciences-humaines.jpg'
+
 
 
 
@@ -34,7 +33,7 @@ export const collections = [
     imageUrl: CarteGeo,
   },
    {
-    title: 'Lettres et Sciences humaines',
+    title: 'Sciences humaines et lettres',
     description: 'Notre bibliothèque possède un large éventail de titres en humanités, en particulier en littérature de langue française. Au sein de nos collections particulières, mentionnons celle de Joseph-Édouard Perrault (4000 documents, dont plusieurs reliés avec goût, incluant certains travaillés par le célèbre relieur français René Kieffer) de Sir Joseph-Édouard Chapleau (4000 ouvrages, dont plusieurs éditions complètes de grands auteurs français) et de Gilles Blain (près de 400 documents consacrés à l’œuvre de Jean Cocteau). En poésie québécoise, les collections de Michel Beaulieu (7000 documents) et de la poétesse Janou Saint-Denis demeurent des incontournables.',
     imageUrl: SciencesHumaines,
   },
@@ -56,7 +55,7 @@ export const collections = [
   {
     title: 'Droit',
     description: 'Surtout composé de sources en droit civil français à la base du Code civil du Bas-Canada de 1866 (l’un des piliers de la société canadienne-française), notre fonds est aussi riche en ouvrages de droit commercial et de droit criminel anglo-saxons.',
-    imageUrl: Droit,
+    imageUrl: SansImage,
   },
   {
     title: 'Sciences politiques, économiques et sociales',
