+++ content/pages/enseignement/index.mdx
@@ -196,42 +196,14 @@ import { DoorOpenIcon, HandshakeIcon, HeartIcon, MountainsIcon, MagnifyingGlassI
         href="https://calendrier.bib.umontreal.ca/space/19189"
         moreText="Réserver à TGD"
       /> 
+
 <CardWithImage
         title="Salle didactique du Campus de Laval"
         Image={DidacthequeLaval}
         href="https://calendrier.bib.umontreal.ca/space/17823"
         moreText="Réserver au Campus Laval"
       /> 
-<CardWithImage
-        title="Salle multifonctionnelle"
-        Image={DidacthequeLaval}
-        href="https://calendrier.bib.umontreal.ca/space/34414"
-        moreText="Réserver au Campus St-Hyacinthe"
-      /> 
-<CardWithImage
-        title="Espace d’innovation et de créativité"
-        Image={DidacthequeLaval}
-        href="https://calendrier.bib.umontreal.ca/space/34410"
-        moreText="Réserver à la bibliothèque mathématiques et informatiques"
-      /> 
-<CardWithImage
-        title="Salle de travail en équipe"
-        Image={DidacthequeLaval}
-        href="https://calendrier.bib.umontreal.ca/space/34412"
-        moreText="Réserver à la bibliothèque mathématiques et informatiques"
-      /> 
-<CardWithImage
-        title="Salle multifonctionnelle"
-        Image={DidacthequeLaval}
-        href="https://calendrier.bib.umontreal.ca/space/34413"
-        moreText="Réserver à la bibliothèque mathématiques et informatiques"
-      /> 
-<CardWithImage
-        title="Salle Perrault"
-        Image={DidacthequeLaval}
-        href="https://calendrier.bib.umontreal.ca/space/34702"
-        moreText="Réserver au secteur Livres rares et collections spéciales"
-      /> 
+
   </Carousel1>
 
 </Section>
