+++ content/pages/index.mdx
@@ -11,7 +11,7 @@ import SuperHero from '@/components/SuperHero'
 import RechercheSofiaAccueil from '@/components/RechercheSofiaAccueil'
 import ListeFormations from '@/components/ListeFormations'
 import ListeEvenements from '@/components/ListeEvenements'
-import ListeNouvellesAccueil from '@/components/ListeNouvellesAccueil'
+import ListeNouvelles from '@/components/ListeNouvelles'
 import LayoutGrid from  '@/components/utils/LayoutGrid'
 
 import { BookmarkSimpleIcon, PersonSimpleCircleIcon, UsersIcon, VideoIcon } from '@phosphor-icons/react'
@@ -48,7 +48,7 @@ import imageEngagement from './images/engagement-accueil.webp'
       <ListeEvenements id="evenements" />
     </Grid>
     <Grid container spacing='2.25rem' size={{ sm: 12, md: 6, lg: 5 }}>
-      <ListeNouvellesAccueil title="Nouvelles" id="nouvelles" />
+      <ListeNouvelles title="Nouvelles" id="nouvelles" />
     </Grid>
   </LayoutGrid>
