+++ content/pages/nouvelles.mdx
@@ -9,10 +9,15 @@ import SuperHero from '@/components/SuperHero'
   imageName="horaire"
 />
 
-import ListeNouvellesCombinees from '@/components/ListeNouvellesCombinees'
+import ListeNouvellesRich from '@/components/ListeNouvellesRich'
+import ListeNouvellesUdeM from '@/components/ListeNouvellesUdeM'
 
 <Section>
 
-  <ListeNouvellesCombinees />
+  <ListeNouvellesRich />
+  
+  <Box sx={{ mt: 6 }}>
+    <ListeNouvellesUdeM title="Nouvelles UdeM sur les bibliothèques" />
+  </Box>
 
 </Section>
