+++ gatsby-node.mjs
@@ -171,83 +171,3 @@ async function doCreateNouvelles({ graphql, actions, reporter }) {
 //     })
 //   }
 // }
-
-/**
- * Ce module gère la récupération et l’intégration des nouvelles externes provenant du flux RSS
- * des nouvelles de l’Université de Montréal (UdeM).
- * La liste des nouvelles se met à jour uniquement lorsqu’on fait un build du site.
- */
-import { parseStringPromise } from 'xml2js'
-import fetch from 'node-fetch'
-
-const fetchUdeMNews = async () => {
-  try {
-    const response = await fetch('https://nouvelles.umontreal.ca/rss/sujets/bibliotheques/')
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`)
-    }
-
-    const xmlText = await response.text()
-    const result = await parseStringPromise(xmlText)
-
-    return result.rss.channel[0].item.map(item => {
-      const description = item.description?.[0] || ''
-      const cdataContent = description.match(/<!\[CDATA\[(.*?)\]\]>/s)?.[1] || description
-
-      return {
-        title: item.title?.[0]?.trim() || 'Sans titre',
-        link: item.link?.[0]?.trim() || '#',
-        description: cdataContent.trim(),
-        pubDate: item.pubDate?.[0] || new Date().toISOString(),
-        enclosure: item.enclosure?.[0]?.$?.url || null,
-        formattedDate: new Date(item.pubDate?.[0] || Date.now()).toLocaleDateString('fr', {
-          year: 'numeric',
-          month: 'long',
-          day: 'numeric'
-        }),
-      }
-    })
-  } catch (error) {
-    console.error('Erreur lors du fetch RSS UdeM:', error)
-    return []
-  }
-}
-
-function formatDate(dateString) {
-  if (!dateString) return ''
-
-  // Si la date est déjà formatée en français (venant de GraphQL)
-  if (dateString.match(/\d{1,2} \w+ \d{4}/)) {
-    return dateString
-  }
-
-  // Pour les dates ISO ou autres formats
-  const date = new Date(dateString)
-  return date.toLocaleDateString('fr', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric'
-  })
-}
-
-export async function sourceNodes({ actions, createNodeId, createContentDigest }) {
-  const { createNode } = actions
-
-  const udeMNews = await fetchUdeMNews()
-
-  udeMNews.forEach((item, index) => {
-    const nodeData = {
-      ...item,
-      type: 'udem',
-    }
-
-    createNode({
-      ...nodeData,
-      id: createNodeId(`udem-news-${index}`),
-      internal: {
-        type: 'UdemNews',
-        contentDigest: createContentDigest(nodeData),
-      },
-    })
-  })
-}
